refactored project1
