+++ src/main/java/edu/project1/Main.java
@@ -1,7 +1,5 @@
 package edu.project1;
 
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
 
 public final class Main {
